Suggestions regarding unicode
I wrote some stuff because company-mode doesn't complete things that are already fully written, meaning that, for example, you can't auto-complete \mu and \nu into unicode. I'm putting in a pull request in case you are interested.

One is an interactive function to replace a fully written unicode character behind the cursor.

The other is an alternate definition of company-math-symbols-unicode. When the sole completion is the word already typed (for example \lambda, but *not* \leftarrow), it adds "--[unicode character]" to the end of the completion string. The reason I did this is so that a completion will execute if you type the whole word and hit the complete key, and thus the conversion will occur. If you want it to be less visible, you could change it to a space or something.