Always offer completions (workaround for company issue #476)
The issue appears clearly when one wants to insert unicode symbols with
short LaTeX names, such as ‘\Cap’, ‘\wp’, or ‘\wr’. The problem is that
company-mode closes its popup if there is only one completion available,
and it has been typed in full. In that case, even running
‘company-manual-begin’ has no effect; the only way to insert the
corresponding unicode symbol is to remove the last character of the
symbol, and to run ‘company-manual-begin’.

This commit introduces the workaround discussed in
https://github.com/company-mode/company-mode/issues/476: we add a space
at the end of each symbol in ‘company-math-symbols-unicode’, to ensure
that the user can always use ‘RET’ to insert Unicode symbols.